requirements.txt

fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2
yfinance==0.2.48
pandas==2.2.3
numpy==2.1.3
httpx==0.27.2
websockets==12.0
sqlmodel==0.0.16
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
Jinja2==3.1.4

